#40007 Lets use int as seconds all the way down

+++ Domain/ApplicationLayer/Rta/Service/Mapping.cs
@@ -21,7 +21,7 @@ namespace Teleopti.Ccc.Domain.ApplicationLayer.Rta.Service
 		public int DisplayColor { get; set; }
 
 		public bool IsAlarm { get; set; }
-		public long ThresholdTime { get; set; }
+		public int ThresholdTime { get; set; }
 		public int AlarmColor { get; set; }
 		
 	}
